Add files via upload
The code uses the CP-SAT solver from the OR-Tools library to solve a graph theory problem. The problem is to find a minimum weight spanning tree in an undirected graph with 21 nodes, where the degree of each node is 5 except for the last node, which has a degree of 6.

The code first defines the number of nodes and their degree using a dictionary called degree, where the key is the node number and the value is its degree.

The code then creates a CP-SAT model object called model and defines three dictionaries X, Y, and Slack.

The X dictionary represents the edges of the graph and is defined using two nested loops that iterate over all pairs of nodes in the graph. For each pair of nodes (i, j), the code creates a new Boolean variable X[(i, j)] and adds it to the model.

The Y dictionary is used to represent auxiliary variables that help to ensure that the selected edges form a tree. For each pair of nodes (i, j) and each node k that is not i or j, the code creates a new Boolean variable Y[(i, j, k)] and adds it to the model.

The Slack dictionary is used to represent a relaxation of the problem constraints. For each pair of nodes (i, j), the code creates a new integer variable Slack[(i, j)] with a domain of [0, 1] and adds it to the model.

The code then defines the objective function to minimize the sum of all Slack variables.

The next set of constraints ensure that each node in the graph has the correct degree. For each node k in the graph, the code adds a new constraint to the model that specifies that the sum of all X variables that involve k must equal its degree.

The next set of constraints ensure that the selected edges form a tree. For each pair of nodes (i, j), the code adds a new constraint to the model that specifies that at least one of X[(i, j)], Y[(i, j, k)], or Slack[(i, j)] must be true.

The final set of constraints ensure that the selected edges do not form cycles. For each pair of nodes (i, j) and each node k that is not i or j, the code adds two new constraints to the model. The first constraint specifies that if Y[(i, j, k)] is true, then either X[(i, k)] or X[(k, i)] must also be true, and the second constraint specifies that if Y[(i, j, k)] is true, then either X[(j, k)] or X[(k, j)] must also be true.

The code then creates a CP-SAT solver object called solver, sets a solver parameter to disable logging of the search progress, and calls the Solve method of the solver object to solve the model.

If the solver returns an optimal solution, the code extracts the values of the X and Y variables and creates a set of edges EDGES that correspond to the selected edges in the graph. Finally, the code prints the values of X, Y, and EDGES.